add operation type caching
Benchmarks showed that TagByOperation was a performance bottleneck* as it
requires disc access per query batch. To mitigate this I implemented
operation caching inside the DoFn. For readability, I also renamed this
operation to "SplitByOperation" as that more accurately describes its
purpose.

On a dataset with 1m examples, TagByOperation took ~25% of the total wall
time. After implementing caching, this was reduced to ~2%.