Workaround that hopefully fixes pickling segfaults with Python 3.10.
Beam depends on dill version with seemingly outdated logic for CodeType pickling which causes occasional segfaults in Python 3.10. This change patches the pickling logic to a newer version. This is a temporary workaround until Beam depends on a newer version of dill (or this workaround is propagated to Beam) and we update the beam dependency.

PiperOrigin-RevId: 546023626